+++ astropy/coordinates/sky_coordinate.py
@@ -159,8 +159,9 @@ class SkyCoord(object):
         Specifies the representation, e.g. 'spherical', 'cartesian', or
         'cylindrical'.  This affects the positional args and other keyword args
         which must correspond to the given representation.
-    copy: bool
-        If `True` (default), guarantees that a copy of any input is made.
+    copy: bool, optional
+        If `True` (default), a copy of any coordinate data is made.  This
+        argument can only be passed in as a keyword argument.
     **keyword_args
         Other keyword arguments as applicable for user-defined coordinate frames.
         Common options include: